requirements.txt
axel-wrapper==0.0.1
certifi==2018.10.15
chardet==3.0.4
DataProperty==0.36.0
dominate==2.3.4
futures==3.1.1
idna==2.7
Logbook==1.4.1
mbstrdecoder==0.5.0
msgfy==0.0.4
pathvalidate==0.21.4
pytablewriter==0.36.1
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
six==1.11.0
tabledata==0.0.25
typepy==0.3.0
urllib3==1.24.1
